[MISC] rename sa_tag_method into split_read_method
